+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2016-XX-XX FreshRSS 1.6.0-dev
 
+* UI
+	* Download icon 💾 for podcasts [#1236](https://github.com/FreshRSS/FreshRSS/issues/1236)
+
 
 ## 2016-08-29 FreshRSS 1.5.0